[Kernels][GPU] Fix the enums for CUBLAS module (#40707)
This fixes the enums sizes for the CUBLAS module. This allows one to run
CUBLAS from within Mojo for testing purposes.

GPU_ORIG_REV_ID: 9d64d9b10d697aa8bb77ace54a24a77eb6fbf010

+++ mojo/stdlib/src/gpu/_cublas/dtype.mojo
@@ -8,7 +8,7 @@
 @value
 @register_passable("trivial")
 struct Property:
-    var _value: Int8
+    var _value: Int32
     alias MAJOR_VERSION = Property(0)
     alias MINOR_VERSION = Property(1)
     alias PATCH_LEVEL = Property(2)
@@ -38,7 +38,7 @@ struct Property:
 @value
 @register_passable("trivial")
 struct DataType:
-    var _value: Int8
+    var _value: Int32
     alias R_16F = DataType(2)
     alias C_16F = DataType(6)
     alias R_16BF = DataType(14)
